Change how DLL files are searched on Windows
The contents of the `_win32_search_file` function is identical to the
default found at <https://metacpan.org/source/BINGOS/ExtUtils-MakeMaker-7.34/lib/ExtUtils/Liblist/Kid.pm#L403>
except the loop switches the `@file_list` and the `@$paths` so that all
possible extensions are found before moving on to the next path.

This can be a problem for example when there are two possible paths such
as:

  /mingw64/lib/site_perl/auto/Cairo/Cairo.a
  /mingw64/lib/libcairo.a

where the first is the built dynamic library for the Cairo.pm module and
the second is the dynamic library for the native cairo library
(e.g., which linked with `pkg-config --libs cairo`).

And the linker arguments are given as:

  g++ [...] \
    -L/mingw64/lib/site_perl/auto/Cairo -lCairo \
    -L/mingw64/lib -lcairo

Both need to be linked, but the path needs to be resolved first before
trying an alternate file name in the file list because the `-L` linker
flag is meant to give that path priority. Otherwise, due to the order of
the file list, `/mingw64/lib/libcairo.a` will be linked twice since
`libcairo.a` will be searched across all paths first.